Keep fractional corner offsets in shape padding
Both padding calculations truncated custom fractional pixel offsets while
the clip path kept them, so content was inset less than the shape was
clipped — a permanent ~1px drift for values like 12.5px.

- Editor: getMaxOffset() used parseInt; use parseFloat to match get-path.js.
- Front end: get_dynamic_shape_padding_value() cast to int; preserve the
  authored fractional value instead. view.js clips the path from the same
  attribute with parseFloat, and nothing re-computes the PHP padding on load.

+++ a8csp-dynamic-shapes/src/js/extend-blocks.js
@@ -104,8 +104,11 @@ const addControls = createHigherOrderComponent( ( BlockEdit ) => {
 								wrapperRef.current
 							);
 						}
-						// Handle regular pixel values.
-						return parseInt( value ) || 0;
+						// Handle regular pixel values. Use parseFloat, not
+						// parseInt, to keep fractional widths (e.g. 12.5px):
+						// get-path.js parses the clip path the same way, so the
+						// preview padding stays aligned with the clipped shape.
+						return parseFloat( value ) || 0;
 					} );
 
 				if ( values.length === 0 ) {
